third_party/WebKit/LayoutTests/external/wpt/cookies/samesite/window-open-reload-expected.txt

This is a testharness.js-based test.
PASS Reloaded same-host auxiliary navigations are strictly same-site.
PASS Reloaded subdomain auxiliary navigations are strictly same-site.
FAIL Reloaded ross-site auxiliary navigations are laxly same-site assert_equals: Non-SameSite cookies are always sent. expected (string) "0.12724065873932022" but got (undefined) undefined
Harness: the test ran to completion.
